Migrated project to gradle.
