+++ src/main/java/com/dzinevich/CatCareApplication.java
@@ -1,4 +1,4 @@
-package org.dzinevich;
+package com.dzinevich;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
