+++ src/main/java/com/dzinevich/controller/CatsController.java
@@ -1,9 +1,16 @@
-package org.dzinevich.controller;
+package com.dzinevich.controller;
 
-import org.dzinevich.model.Cat;
-import org.dzinevich.repository.CatRepository;
+import com.dzinevich.model.Cat;
+import com.dzinevich.repository.CatRepository;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 import java.net.URI;
 import java.net.URISyntaxException;
