+++ src/test/java/com/dzinevich/CatCareApplicationTest.java
@@ -1,0 +1,14 @@
+package com.dzinevich;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CatCareApplicationTest {
+
+    @SuppressWarnings("PMD.JUnitTestsShouldIncludeAssert")
+    @Test
+    void main() {
+        //ApplicationContext is loading successfully
+    }
+}
